requirements.txt
fastapi[standard]
vnstock
pandas
python-dotenv
pytz
cachetools